Create README.md
The aim of this project was to create a home automation system with the use of Raspberry Pi. Home Automation is a general term for making a house, apartment or any other property ”smart”, so that appliances can be controlled remotely. Moreover they can be automated, for example lights switching on when user wakes up or coffee machine getting switched on from a phone while user is coming home.

For a project to be finished certain conditions had to be met such as hardware components being built and comprehensive software to control appliances being developed. Overall this project was about creating a cheap and easy to use system that anyone could afford and then customize to their needs.

Technologies used :

Hardware : 
- Raspberry Pi 2 
- Radio Frequency Identification ( RFID )sensor EM18 
- Temperature sensor 18b20
- Passive Infrared ( PIR ) motion sensor
- Moisture sensor 
- 12V motor
- Motor driver L294D

Python hardware coding 

Backend : 
- mysql